# Use ffjson for encoding/decoding.