--- a/Godeps/_workspace/src/github.com/pquerna/ffjson/.gitignore
+++ b/Godeps/_workspace/src/github.com/pquerna/ffjson/.gitignore
@@ -1,0 +1,4 @@
+tests/go.stripe/ff/customer_ffjson.go
+tests/goser/ff/goser_ffjson.go
+tests/types/ff/everything_ffjson.go
+tests/ff_ffjson.go