--- a/Godeps/_workspace/src/github.com/pquerna/ffjson/NOTICE
+++ b/Godeps/_workspace/src/github.com/pquerna/ffjson/NOTICE
@@ -1,0 +1,8 @@
+ffjson
+Copyright (c) 2014, Paul Querna
+
+This product includes software developed by 
+Paul Querna (http://paul.querna.org/).
+
+Portions of this software were developed as
+part of Go, Copyright (c) 2012 The Go Authors.