… Godeps/Godeps.json
… Godeps/_workspace/src/github.com/pquerna/ffjson/.travis.yml
… Godeps/_workspace/src/github.com/pquerna/ffjson/README.md
… Godeps/_workspace/src/github.com/pquerna/ffjson/fflib/v1/buffer.go
… Godeps/_workspace/src/github.com/pquerna/ffjson/fflib/v1/bytenum.go
… Godeps/_workspace/src/github.com/pquerna/ffjson/fflib/v1/decimal.go
… Godeps/_workspace/src/github.com/pquerna/ffjson/fflib/v1/extfloat.go
… Godeps/_workspace/src/github.com/pquerna/ffjson/fflib/v1/fold.go
… Godeps/_workspace/src/github.com/pquerna/ffjson/fflib/v1/ftoa.go
… Godeps/_workspace/src/github.com/pquerna/ffjson/fflib/v1/internal/atof.go
… Godeps/_workspace/src/github.com/pquerna/ffjson/fflib/v1/internal/atoi.go
… Godeps/_workspace/src/github.com/pquerna/ffjson/fflib/v1/internal/extfloat.go
… Godeps/_workspace/src/github.com/pquerna/ffjson/fflib/v1/internal/ftoa.go
… Godeps/_workspace/src/github.com/pquerna/ffjson/fflib/v1/iota.go
… Godeps/_workspace/src/github.com/pquerna/ffjson/fflib/v1/jsonstring.go
… Godeps/_workspace/src/github.com/pquerna/ffjson/fflib/v1/jsonstring_test.go
… Godeps/_workspace/src/github.com/pquerna/ffjson/fflib/v1/lexer.go
… Godeps/_workspace/src/github.com/pquerna/ffjson/fflib/v1/lexer_test.go
… Godeps/_workspace/src/github.com/pquerna/ffjson/fflib/v1/reader.go
… Godeps/_workspace/src/github.com/pquerna/ffjson/fflib/v1/reader_scan_amd64.go
… Godeps/_workspace/src/github.com/pquerna/ffjson/fflib/v1/reader_scan_amd64.s
… Godeps/_workspace/src/github.com/pquerna/ffjson/fflib/v1/reader_scan_generic.go
… Godeps/_workspace/src/github.com/pquerna/ffjson/fflib/v1/reader_test.go
… Godeps/_workspace/src/github.com/pquerna/ffjson/generator/generator.go
… Godeps/_workspace/src/github.com/pquerna/ffjson/generator/inceptionmain.go
… Godeps/_workspace/src/github.com/pquerna/ffjson/generator/parser.go
… Godeps/_workspace/src/github.com/pquerna/ffjson/generator/tags.go
… Godeps/_workspace/src/github.com/pquerna/ffjson/generator/tempfile.go
… Godeps/_workspace/src/github.com/pquerna/ffjson/inception/decoder.go
… Godeps/_workspace/src/github.com/pquerna/ffjson/inception/decoder_tpl.go