docs indicate 2-3x faster for encode/decode of JSON

@wojtek-t @fgrzadkowski @davidopp