@brendandburns 2015-06-05
fixed ```gofmt``` problems.