@bgrant0607 2015-06-05
Before/after perf test results?

verify-boilerplate.sh failed. Do you have commit hooks installed?