@lavalamp 2015-06-06
yeah the autogenerated files need the boilerplate headers.