@smarterclayton 2015-06-06
I love me some before and after %s

> On Jun 5, 2015, at 7:55 PM, Brian Grant <notifications@github.com> wrote:
> 
> Before/after perf test results?
> 
> —
> Reply to this email directly or view it on GitHub.
>