@philips 2015-06-06
Can we choose one code generating JSON library? @fgrzadkowski just got something using code into go-etcd: https://github.com/coreos/go-etcd/pull/218