@brendandburns 2015-06-06
I'll check out ugorji.

@smarterclayton Numbers for you (tl; dr; closer to a 1.5x speed up vs the 2-3x promised...)

fwiw, I looked into the generated code, it is the ```map[string]string``` for labels that is still getting kicked into the std. json library.

#### Encode
Before:
```
$ go test -bench=BenchmarkEncode -run=XX github.com/GoogleCloudPlatform/kubernetes/pkg/api --benchtime=5s --benchmem
PASS
BenchmarkEncode	  200000	     57875 ns/op	   11936 B/op	     112 allocs/op
BenchmarkEncodeJSON	  200000	     34788 ns/op	    7312 B/op	      40 allocs/op
```

After:
```
$ go test -bench=BenchmarkEncode -run=XX github.com/GoogleCloudPlatform/kubernetes/pkg/api --benchtime=5s --benchmem
PASS
BenchmarkEncode	  200000	     41558 ns/op	    7128 B/op	     126 allocs/op
BenchmarkEncodeJSON	  500000	     17956 ns/op	    1967 B/op	      53 allocs/op
```

#### Decode
Before:
```
$ go test -bench=BenchmarkDecode -run=XX github.com/GoogleCloudPlatform/kubernetes/pkg/api --benchtime=5s --benchmem
PASS
BenchmarkDecode	   50000	    123256 ns/op	   10850 B/op	     182 allocs/op
BenchmarkDecodeInto	  100000	    123268 ns/op	   10818 B/op	     182 allocs/op
BenchmarkDecodeJSON	  100000	     68811 ns/op	    5121 B/op	      83 allocs/op
```

After:
```
$ go test -bench=BenchmarkDecode -run=XX github.com/GoogleCloudPlatform/kubernetes/pkg/api --benchtime=5s --benchmem
PASS
BenchmarkDecode	  100000	     97864 ns/op	    9898 B/op	     168 allocs/op
BenchmarkDecodeInto	  100000	     93739 ns/op	    9898 B/op	     168 allocs/op
BenchmarkDecodeJSON	  200000	     34841 ns/op	    4164 B/op	      69 allocs/op
```