@brendandburns 2015-06-10
Closing this for now

I'll re-open when it is ready for consideration.